AbilityGenerator working with Ant, along with newly broken up Ability classes now moving to their own folder
