+++ pet.py
@@ -11,7 +11,7 @@ class Pet:
             3: ability3
         }
         self.abilities = {
-            lvl: ability_generator(ability_dict).generate()
+            lvl: ability_generator(ability_dict, self).generate()
             for lvl, ability_dict in self.ability_dicts.items()
         }
         self.ability = self.abilities[self.level]
@@ -43,9 +43,9 @@ class Pet:
 
         if not self.is_alive() or not enemy_pet.is_alive():
             if not self.is_alive():
-                self.ability.trigger("faint", self, self.team, enemy_team=enemy_pet.team)
+                self.ability.trigger("Faint", self, self.team, enemy_team=enemy_pet.team)
             if not enemy_pet.is_alive():
-                enemy_pet.ability.trigger("faint", enemy_pet, enemy_pet.team, enemy_team=self.team)
+                enemy_pet.ability.trigger("Faint", enemy_pet, enemy_pet.team, enemy_team=self.team)
 
             # Clean up dead pets after abilities have been triggered
             if not self.is_alive() and self in self.team.pets:
@@ -54,10 +54,11 @@ class Pet:
                 enemy_pet.team.remove_pet(enemy_pet)
 
     def faint(self):
+        print("Faint")
         if not self.fainted:
             self.fainted = True
             if self.ability:
-                self.ability.trigger("faint", self, self.team)
+                self.ability.trigger("Faint", self, self.team)
 
     def apply_ability(self, team, enemy_team):
         self.ability.apply(self, team, enemy_team)
